+++ src/components/BetaPage/LinkBack.jsx
@@ -3,7 +3,7 @@ import ArrowBackIosIcon from "@mui/icons-material/ArrowBackIos";
 import { Link } from "react-router-dom";
 
 export const LinkBack = ({ to }) => (
-  <IconButton component={Link} to={to} size="large">
+  <IconButton component={Link} to={to} size="small">
     <ArrowBackIosIcon fontSize="inherit" />
   </IconButton>
 );