+++ src/pages/BetaAccount.jsx
@@ -7,16 +7,20 @@ import {
   ListItemText,
   ListItemAvatar,
   ListSubheader,
-  Avatar
+  Avatar,
+  Fab
 } from "@mui/material";
 import AddIcon from "@mui/icons-material/Add";
 
 import WorkIcon from "@mui/icons-material/Work";
 import { Link } from "react-router-dom";
 
+import { useScreenSize } from "../utils/use-screen-size";
 import BetaPage from "../components/BetaPage";
 
 const BetaAccount = () => {
+  const { isBigScreen, isSmallScreen } = useScreenSize();
+
   const { uname, accountData } = useBetaAccount();
 
   const { own_projects: ownProjects, shared_projects: sharedProjects } =
@@ -26,14 +30,16 @@ const BetaAccount = () => {
     <BetaPage
       title="Hello, Unknown User"
       actions={
-        <Button
-          component={Link}
-          to={`/beta/${uname}/create`}
-          variant="contained"
-          startIcon={<AddIcon />}
-        >
-          New&nbsp;project
-        </Button>
+        isBigScreen && (
+          <Button
+            component={Link}
+            to={`/beta/${uname}/create`}
+            variant="contained"
+            startIcon={<AddIcon />}
+          >
+            New&nbsp;project
+          </Button>
+        )
       }
     >
       <Paper>
@@ -78,6 +84,23 @@ const BetaAccount = () => {
           ))}
         </List>
       </Paper>
+      {isSmallScreen && (
+        <Fab
+          component={Link}
+          to={`/beta/${uname}/create`}
+          color={"primary"}
+          sx={{
+            margin: 0,
+            top: "auto",
+            right: 30,
+            bottom: 30,
+            left: "auto",
+            position: "fixed"
+          }}
+        >
+          <AddIcon />
+        </Fab>
+      )}
     </BetaPage>
   );
 };